feat(match): 에러 코드 정리
